collection: Add SharedData type
A common pattern I encounter is where someone sends a message and waits
for a reply. The response message is usually returned from another task
or socket (scheduled on the event loop) and my regular approach is to
synchronize with a semaphore. This new structure, SharedData, uses an
asyncio Event and allows someone to wait for data and someone else to
post it:

  shared = SharedData[int]()

  # In one task
  resp: int = await shared.wait(timeout=5)

  # Another task
  shared.set(123)

I will try to migrate to this structure in existing protocols over time.

Relates to #1192